fix: eliminate circular import between types.ts and game-state.ts
Move PreMatchState interface to types.ts (where GameState lives) to
avoid the circular `import type` that could break Next.js/webpack
builds in CI. Inline PurchasedInducement shape to keep types.ts
import-free. Re-export PreMatchState from game-state.ts for
backward compatibility.
